requirements.txt
streamlit
openai
langchain
faiss-cpu
pdfplumber
python-dotenv
numpy